requirements.txt
#Se reqwuieren instalar las siguientes dependencias
fastapi
uvicorn
psycopg2-binary
python-dotenv
rasterio
python-multipart
shapely